Disconnection synchronization
When we destroy transport connection, we simply remove uid-map in
tincan. Potentially, if one node destroy connection, the remote
node still can consider the connection exists leading to unexpected
behaviros such as forwarding packet to disconnected node or
destroying connection right after the connection is re-established
when previous connection is destroyed in a near past.